Pipewire backend

+++ Opcodes/serial.c
@@ -405,7 +405,7 @@ int32_t serialWrite_S(CSOUND *csound, SERIALWRITE *p)
     if (UNLIKELY(write((int32_t)*p->port,
                        ((STRINGDAT*)p->toWrite)->data,
                        ((STRINGDAT*)p->toWrite)->size))!=
-        ((STRINGDAT*)p->toWrite)->size) /* Does Windows write behave correctly? */
+        (int64_t) ((STRINGDAT*)p->toWrite)->size) /* Does Windows write behave correctly? */
         return NOTOK;
 #else
       int32_t nbytes;
